Moved lm75 temperature reading to the rev10 board, since it is a board specific sensor
