+++ src/system-plugins/platform-manager/platforms/raspberrypi/boards/rev10/setup.js
@@ -1,9 +1,19 @@
-const fs            = require('fs');
+const Promise       = require( 'bluebird' );
+const fs            = Promise.promisifyAll( require('fs') );
 const path          = require('path');
 const spawn         = require('child_process').spawn;
 const SerialBridge  = require('TridentSerialBridge.js');
+var Periodic        = require( 'Periodic' );
 
 var debug = {};
+
+// I2C setup
+var i2c       = require('i2c');
+var lm75Address = 0x48;
+var i2c1  = new i2c( lm75Address, { device: '/dev/i2c-1' } );
+
+// Promisify the I2C read operation
+var i2c1ReadAsync = Promise.promisify( i2c1.read, { context: i2c1 } );
 
 var SetupBoardInterface = function(board) 
 {
@@ -21,6 +31,26 @@ var SetupBoardInterface = function(board)
     // ------------------------------------------------
     // Setup private board methods
     // ------------------------------------------------
+
+    // Create periodic function for reading LM75 temperature (underneath the RPI)
+    board.readLM75Temp = new Periodic( 1000, "timeout", () =>
+    {
+      // Read two bytes to get the temperature bits
+      return i2c1ReadAsync( 2 )
+        .then( (result) =>
+        {
+          // First byte is temperature in C, first bit of second byte is equal to 0.5C
+          let temp = result[ 0 ] + ( ( ( result[ 1 ] & 0x80 ) >> 7 ) * 0.5 );
+
+          // Emit on cockpit bus
+          board.cockpit.emit( "board.temp.lm75", temp );
+        })
+        .catch( (err) =>
+        {
+          console.log( "Error reading LM75. Stopping task: " + err.message );
+          board.readLM75Temp.stop();
+        });
+    });
 
     // Setup bridge interface event handlers
     board.bridge.on('serial-recieved', function(data) 
@@ -48,6 +78,9 @@ var SetupBoardInterface = function(board)
     // Create and start statemachine
     board.fsm = require( './statemachine.js' )( board );
     board.fsm._e_init();
+
+    // Start periodic tasks
+    board.readLM75Temp.start();
 
     console.log( "Done" );
 };
