+++ src/system-plugins/platform-manager/platforms/raspberrypi/cpu/setup.js
@@ -4,40 +4,12 @@ var path      = require('path');
 var Periodic  = require( 'Periodic' );
 var debug = {};
 
-// I2C setup
-var i2c       = require('i2c');
-var lm75Address = 0x48;
-var i2c1  = new i2c( lm75Address, { device: '/dev/i2c-1' } );
-
-// Promisify the I2C read operation
-var i2c1ReadAsync = Promise.promisify( i2c1.read, { context: i2c1 } );
-
 var SetupCPUInterface = function( cpu ) 
 {
   debug = cpu.debug;
 
   // Decorate the CPU interface with cpu specific properties
   cpu.stats = {};
-
-  // Create periodic functions for reading CPU temp and LM75 temp
-  cpu.readLM75Temp = new Periodic( 1000, "timeout", () =>
-  {
-    // Read two bytes to get the temperature bits
-    return i2c1ReadAsync( 2 )
-      .then( (result) =>
-      {
-        // First byte is temperature in C, first bit of second byte is equal to 0.5C
-        let temp = result[ 0 ] + ( ( ( result[ 1 ] & 0x80 ) >> 7 ) * 0.5 );
-
-        // Emit on cockpit bus
-        cpu.cockpit.emit( "cpu.temp.lm75", temp );
-      })
-      .catch( (err) =>
-      {
-        console.log( "Error reading LM75. Stopping task: " + err.message );
-        cpu.readLM75Temp.stop();
-      });
-  });
 
   cpu.readCPUTemp = new Periodic( 1000, "timeout", () =>
   {
@@ -58,7 +30,6 @@ var SetupCPUInterface = function( cpu )
   });
 
   // Start periodic tasks
-  cpu.readLM75Temp.start();
   cpu.readCPUTemp.start();
 
   // ------------------------------------------------
